requirements.txt
psycopg2-binary
scikit-learn
pandas
numpy
flask
matplotlib
nltk


